requirements.txt
Django==3.1.7
django-rest-swagger==2.2.0
ipython==7.21.0
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-cors-headers==3.7.0
python-dotenv==0.17.0
sqlparse==0.4.1
pytz==2021.1
asgiref==3.3.4
Pillow==8.2.0